Implement new testing suite using Karma - integrates Karma, Webpack, TypeScript, Jasmine, Imagediff - offers legacy JS support via Webpack's script-loader - runs in Chrome or Chrome Headless - removes imagediff @types - switches to forked version of imagediff - switches to forked version of karma-jasmine-jquery - removes remnants of old testing suite
