--- kive/SpecRunner.html
@@ -1,34 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-  <meta charset="utf-8">
-  <title>Kive Spec Runner - Jasmine v2.3.4</title>
-  <!-- Open this in a browser to run all the Javascript unit tests. -->
-
-  <link rel="shortcut icon" type="image/png" href="portal/js/jasmine-2.3.4/jasmine_favicon.png">
-  <link rel="stylesheet" href="portal/js/jasmine-2.3.4/jasmine.css">
-
-  <script src="portal/js/jasmine-2.3.4/jasmine.js"></script>
-  <script src="portal/js/jasmine-2.3.4/jasmine-html.js"></script>
-  <script src="portal/js/jasmine-2.3.4/boot.js"></script>
-  <script src="portal/js/imagediff.js"></script>
-
-  <!-- include source files here... -->
-  <script src="portal/static/portal/jquery-2.0.3.min.js"></script>
-  <script src="portal/static/portal/noxss.js"></script>
-  <script src="portal/static/portal/permissions.js"></script>
-  <script src="librarian/static/librarian/md5.js"></script>
-  <script src="sandbox/static/sandbox/choose_inputs.js"></script>
-  <!--   Move to async module tests   -->
-  <!--<script src="static/sandbox/PipelineRunTable.js"></script>-->
-
-  <!-- include test files here... -->
-  <script src="portal/tests.js"></script>
-  <script src="librarian/tests.js"></script>
-  <script src="sandbox/tests.js"></script>
-
-</head>
-
-<body>
-</body>
-</html>
