+++ kive/pipeline/static/pipeline/pipeline_dialogs.ts
@@ -1,7 +1,7 @@
 
 import { RawNode, CdtNode, MethodNode, OutputNode } from "./canvas/drydock_objects";
 import { CanvasState } from "./canvas/drydock";
-import 'jquery';
+import * as $ from 'jquery';
 
 /**
  * Mini jQuery plugin to make dialogs draggable.
