+++ kive/pipeline/test_assets/pipeline/magnet-highlight-new-fill.png
