+++ kive/pipeline/test_assets/pipeline/magnet-highlight.png
