+++ kive/pipeline/test_assets/pipeline/magnet-unlit.png
