+++ kive/pipeline/test_assets/pipeline/outputnode.png
