+++ kive/pipeline/test_assets/pipeline/rawnode.png
