+++ kive/pipeline/test_assets/pipeline/step.png
