+++ kive/pipeline/tests/contextmenu.tests.ts
@@ -1,12 +1,7 @@
-"use strict";
-
 import { CanvasState } from "../static/pipeline/canvas/drydock";
 import { CanvasContextMenu } from "../static/pipeline/canvas/drydock_contextmenu";
 import { RawNode } from "../static/pipeline/canvas/drydock_objects";
-import "jasmine";
-import 'jquery';
-import 'jasmine-jquery';
-import 'imagediff';
+import * as imagediff from 'imagediff';
 
 describe("Drydock context menu", function() {
     let testMenu;
