+++ kive/sandbox/static/sandbox/PipelineRunTable.ts
@@ -1,8 +1,9 @@
 "use strict";
 
-import { CanvasState, Pipeline } from "static/pipeline/pipeline_all";
-import 'jquery';
-declare var permissions: any;
+import { CanvasState, Pipeline } from "../../../pipeline/static/pipeline/pipeline_all";
+(window as any).$ = $;
+require("script-loader!../../../portal/static/portal/permissions.js");
+declare var permissions: { [key: string]: any };
 
 permissions.PipelineRunTable = function($table, is_user_admin, $navigation_links) {
     permissions.PermissionsTable.call(this, $table, is_user_admin, $navigation_links);
