+++ kive/tests.ts
@@ -1,0 +1,4 @@
+import './portal/portal.tests';
+import './sandbox/sandbox.tests';
+import './librarian/librarian.tests';
+import './pipeline/pipeline.tests';
