+++ package.json
@@ -2,22 +2,25 @@
   "name": "kive",
   "version": "0.8.4",
   "devDependencies": {
+    "@metahub/karma-jasmine-jquery": "^1.1.1",
     "@types/jasmine": "^2.5.35",
     "@types/jasmine-ajax": "^3.1.31",
     "@types/jasmine-jquery": "^1.5.28",
-    "browser-sync": "^2.17.3",
     "grunt-contrib-jshint": "^1.0.0",
     "grunt-tslint": "^3.2.1",
-    "imagediff": "^1.0.8",
-    "jasmine": "^2.8.0",
+    "imagediff": "jamesnakagawa/js-imagediff.git#upgrade-canvas-and-jasmine",
     "jasmine-ajax": "^3.2.0",
+    "jasmine-core": "2.5.2",
     "jasmine-jquery": "^2.1.1",
     "jshint": "^2.9.4",
     "karma": "^1.7.1",
     "karma-chrome-launcher": "^2.2.0",
     "karma-jasmine": "^1.1.0",
+    "karma-jasmine-ajax": "^0.1.13",
+    "karma-jasmine-html-reporter": "naganowl/karma-jasmine-html-reporter#a89c5e2eff1b332dcb857b4d01eee6a63ccee229",
+    "karma-sourcemap-loader": "^0.3.7",
+    "karma-spec-reporter": "0.0.31",
     "karma-webpack": "^2.0.4",
-    "puppeteer": "0.10.2",
     "shelljs": "^0.7.5",
     "tslint": "^3.15.1"
   },
@@ -25,29 +28,28 @@
     "@types/jquery": "^2.0.33",
     "@types/node": "^6.0.45",
     "awesome-typescript-loader": "^3.2.3",
+    "script-loader": "^0.7.1",
     "core-js": "^2.4.1",
     "css-loader": "^0.28.7",
     "grunt": "^1.0.1",
     "grunt-cli": "^1.2.0",
     "grunt-regex-replace": "^0.4.0",
     "jquery": "^3.1.1",
-    "node-sass": "^3.10.1",
+    "node-sass": "^4.0.0",
     "sass-loader": "^6.0.6",
     "source-map-loader": "^0.2.1",
     "style-loader": "^0.18.2",
-    "systemjs": "^0.19.39",
     "typescript": "^2.0.3",
     "uglifyjs-webpack-plugin": "^0.4.6",
     "webpack": "^3.5.6"
   },
   "scripts": {
-    "install": "npm run install-imagediff-@types && grunt pngicons && tsc",
-    "install-imagediff-@types": "mkdir node_modules/@types/imagediff 2> /dev/null; cp kive/portal/js/imagediff.d.ts node_modules/@types/imagediff/index.d.ts",
+    "install": "grunt pngicons && tsc",
     "watch-all": "npm run ts-watch & npm run sass-watch",
     "ts-watch": "tsc -w",
     "sass-watch": "node-sass --watch --recursive --output kive --source-map true --source-map-contents kive",
-    "test:browser": "karma start --browsers Chrome karma.conf.js",
-    "test": "karma start --single-run --browsers ChromeHeadless karma.conf.js"
+    "test:browser": "karma start --browsers Chrome",
+    "test": "karma start --browsers ChromeHeadless"
   },
   "description": "Kive ====",
   "repository": {
@@ -77,8 +79,8 @@
     },
     {
       "name": "James Nakagawa",
-      "email": "james@nakagawa.ca",
-      "url": ""
+      "email": "jamesnakagawa@cfenet.ubc.ca",
+      "url": "http://nakagawa.ca"
     },
     {
       "name": "T. Nguyen",
